[OWL-151] disable authentication for /api/endpoints
